+++ packages/js-client-rest/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @qdrant/js-client-rest
 
+## 1.6.0
+
+### Minor Changes
+
+-   Qdrant v1.6.0 API
+
 ## 1.5.0
 
 ### Minor Changes